Feat: 11.21 예약 reserve 메서드 수정 (itemSchedule 목록 가져오기)

+++ customer/src/main/java/com/triplan/service/MemberServiceImpl.java
@@ -82,7 +82,7 @@ public class MemberServiceImpl implements MemberService {
         MemberVO memberVO = new MemberVO();
         memberVO.setMemberId(memberId);
 
-        if (!files.isEmpty()) {
+        if (files != null && !files.isEmpty()) {
             List<AttachmentVO> filesToDelete = attachmentMapper.select(AboutTableType.MEMBER, memberId);
             AttachmentUtil.deleteAttachments(filesToDelete);
             attachmentMapper.delete(AboutTableType.MEMBER, memberId);
